requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-htmx>=1.17
python-dotenv>=1.0
